+++ src/booleano/nodes/converters.py
@@ -22,7 +22,7 @@ evaluable ones).
 """
 
 from booleano.exc import ConversionError
-from booleano.nodes.operands import String, Number, Set
+from booleano.nodes.constants import String, Number, Set
 from booleano.nodes.operations import (Not, And, Or, Xor, Equal, NotEqual,
     LessThan, GreaterThan, LessEqual, GreaterEqual, BelongsTo, IsSubset,
     UnaryOperation)