+++ tests/parsing/test_trees.py
@@ -24,7 +24,7 @@ from nose.tools import eq_, ok_, assert_false, assert_raises
 
 from booleano.parser.trees import EvaluableParseTree, ConvertibleParseTree
 from booleano.nodes.operations import And
-from booleano.nodes.operands import String, PlaceholderVariable
+from booleano.nodes.constants import String, PlaceholderVariable
 from booleano.exc import InvalidOperationError
 
 from tests.utils.mock_converters import AntiConverter