+++ tests/utils/mock_converters.py
@@ -22,7 +22,7 @@ Mock converters.
 from booleano.nodes.converters import BaseConverter
 from booleano.nodes.operations import (Not, And, Or, Xor, Equal, NotEqual,
     LessThan, GreaterThan, LessEqual, GreaterEqual, BelongsTo, IsSubset)
-from booleano.nodes.operands import (String, Number, Set, PlaceholderVariable,
+from booleano.nodes.constants import (String, Number, Set, PlaceholderVariable,
                                      PlaceholderFunction)